+++ test/selectors/date_picker_test.dart
@@ -30,8 +30,8 @@ void main() {
           ),
         );
 
-        expect(find.text('${today.day}'), findsOneWidget);
         expect(find.text('/'), findsNWidgets(2));
+        expect(find.text('${today.day}'), findsOneWidget);
         expect(find.text('${today.month}'), findsOneWidget);
         expect(find.text('${today.year}'), findsOneWidget);
       },